Use genalloc to allocate SRAM
In newer kernel, SRAM is usually managed by the genalloc infra-
structure. Use the genalloc API to allocate the SRAM are used
for MCC bookkeeping data.

When the user requests to load a Firmware in SRAM, we would need
to do the same too. However, this currently conflicts with PM
(suspend/resume) anyway. For now, still allow loading into SRAM
without allocating it properly, with an appropriate warning.

We might later disallow using SRAM for the M4 when suspend/resume
is enabled, or similar measures to prevent this conflict.